+++ src/main/java/com/api/bank/dto/EventRequest.java
@@ -1,4 +1,4 @@
 package com.api.bank.dto;
 
-public record EventRequest(String type, Integer origin, Integer amount, Integer destination) {
+public record EventRequest(String type, Integer origin, Double amount, Integer destination) {
 }
